requirements.txt
# Core
discord.py>=2.3,<3
aiohttp>=3.9,<4
pyyaml>=6.0
